Make CSP 'nonce' available to GTM as 'data-nonce'
Our performance analyst needs to be able to assign the request CSP
'nonce' value as a variable in GTM so the GTM custom tags can use this
'nonce'.
This way the will be able to run once injected by GTM into the
user browsers without being blocked by the Content Security Policy.

But when loading in Chrome and in GTM, the 'nonce' value seems stripped
out and loads as empty string ''

Following some online guides suggestions, we pass it also as data
attribute (data-nonce) and seems to load the value in the script
attribute in Chrome, what is a good sign.

Pushing it to GTM so variables can be set from it.