Exclude main deploys from cancelling review app deletions
Noticed a lot of review app deletion cancelled jobs because their
reference resolution turned to be "main" and the "main" branch
deployment caused the concurrency lock to be matched and the deletion
job ended cancelled. This turns into a dangling review app staying
forever unless manually deleted.

This attempt does:
- Try to use the PR number.
- If not try the reference, as long as is not "main"
- If it is "main", sets a unique id per job for the concurrency.