Feature: Steps form can define other steps to reset
This feature enables a form object to declare a method that lists which
other steps will be removed from the vacancy completed steps list upon
form submission.

Set in the form as allows us to do "if this data is in the form, this
form mark this other step/s as incomplete again".

The practical purpose on this case will be:
- When contact details for an already existent/filled vacancy get
  updated, if the new contact email does not belong to a registered
  publisher email, we can removed the "confirm_contact_details" step
  from the list of completed steps in the vacancy.
- This will enable us to show/hide/validate the confirm_contact_details
  later based in "it is completed or not?" criteria.