Contact details flow tweaks proposal (#8238)
## Trello card URL
Refactor proposal on some bits of
https://trello.com/c/f5SGqquV/2132-contact-details-tweaks-in-job-listing-journey-redirect-application-notifications-la-notifications-issue

## Changes in this PR:

### TLDR
Relies in removing/checking the `confirm_contact_details` step from the
vacancy completed steps to make the `confirm_contact_details_form`
behave as a "standard form" requiring less custom code for this step
injected around.

### General idea
The main idea behind this is to try to keep the generic part of our
wizard codebase as generic as possible, reducing the amount of custom
code we had to introduce for the confirm contact details step's "odd"
behaviour (compared to the standard step's behaviours).

Also tries to stick more to `Wicked::Wizard` conventions when possible,
with things like "skip the step if needed" done as indicated by the
Wicked::Wizard docs (with a skip clause on the #show).

### What is a pain with the new Confirm step
- Depends on the previous contact_details step value to be
displayed/required or not.
- The step answer is ephemeral. It doesn't get recorded anywhere.

The new step is, effectively, a "Confirmation/Agreement" page that may
or may not need the agreement and doesn't get recorded anywhere.

### How this PR attempts to make it a bit less painful
- Declares the new step as a permanent member in the Wicked::Wizard list
of steps.
- Skips it when it is not necessary.
- Adds the step to the list of steps to validate before vacancy
submission.
- Uses the `Vacancy#completed_steps` as criteria to check if the
confirmation has been given or not for the contact email. We may not
record the answer, but we record if we completed the step.
- **The important bit: Tries to change the logic in the specific
form/step logic rather than having to inject individual cases in the
general redirection logic. The whole aim is "keep the generic bits as
generic as possible".**

As an output:
- When building a new vacancy:
  - If the contact email belongs to a publisher: 
    - The confirm step is skipped.
  - If the contact email does not belong to a publisher:
- The step is displayed and only added to `Vacancy#completed_steps` if
the user answered "Yes" to the confirmation section.
- The vacancy draft will be submittable as the new form, even in the
list of validatable steps for publishing the vacancy, will not fail
validation unless the step is not listed as completed for the vacancy.

- When updating an already filled vacancy:
  - If the new contact email belongs to a publisher:
- Do nothing. The confirmation page will be skipped in the flow and pass
validation when inquired.
  -  If the new contact email does not belong to a publisher:
- **Set the confirmation step for removal. When the vacancy gets updated
with the form data, the following confirmation step is removed from the
vacancy completed steps.**
- **Since the confirmation step is not listed in the vacancy completed
steps anymore, it is required, displayed for the user and will need to
be completed to make the vacancy "valid" again.**

## Screenshots of UI changes:

### Before

### After


## Checklists:

### Data & Schema Changes

If this PR modifies data structures or validations, check the following:

- [ ] Adds/removes model validations
- [ ] Adds/removes database fields
- [ ] Modifies Vacancy enumerables (phases, working patterns, job roles,
key stages, etc.)

<details>
<summary>If any of the above options has changed then the author must
check/resolve all of the following...</summary>

### Integration Impact

Does this change affect any of these integrations?
- [ ] DfE Analytics platform
- [ ] Legacy imports mappings
- [ ] DWP Find a Job export mappings
- [ ] Publisher ATS API (may require mapping updates or API versioning)

### User Experience & Data Integrity

Could this change impact:
- [ ] Existing subscription alerts (will legacy subscription search
filters break?)
- [ ] Legacy vacancy copying (will copied vacancies fail new
validations?)
- [ ] In-progress drafts for Vacancies or Job Applications
</details>

---------

Co-authored-by: Kyle MacPherson <kaimac2010@gmail.com>