Add validation to interview dates and order applications on offer and… (#8224)
This PR adds date validation to ensure offer and interview feedback
dates are after the interview date, and ensures declined date is after
offer date. It also sorts the job applications by the interviewing_at
date on the interview page and offer_at on the offer page, with
applications with blank values in this field appearing first.