Add tracking & testing for unknown organisation category errors
When Publishers dfe sign-in requests come with the wrong organisation
category (not explicitly accepted by our service), the publishers get
redirected to an error page informing them about this.

Since this has recently coming with the wrong values and we weren't
aware of it until the publishers started opening support requests, we
want to improve this area by pushing this errors (as other DfE Sign-in
errors are) to Sentry.

This will allow us to track if:
- This has been resolved.
- If it starts happening again.