Add tracking & testing for unknown organisation category errors (#8261)
## Trello card URL
-
https://trello.com/c/mzTeEHJF/2350-support-query-mat-access-being-rejected-as-pims-training-provider

## Changes in this PR:

When Publishers dfe sign-in requests come with the wrong organisation
category (not explicitly accepted by our service), the publishers get
redirected to an error page informing them about this.

Since this has recently coming with the wrong values and we weren't
aware of it until the publishers started opening support requests, we
want to improve this area by pushing this errors (as other DfE Sign-in
errors are) to Sentry.

This will allow us to track if:
- This has been resolved.
- If it starts happening again

## Screenshots of UI changes:

### Before

### After


## Checklists:

### Data & Schema Changes

If this PR modifies data structures or validations, check the following:

- [ ] Adds/removes model validations
- [ ] Adds/removes database fields
- [ ] Modifies Vacancy enumerables (phases, working patterns, job roles,
key stages, etc.)

<details>
<summary>If any of the above options has changed then the author must
check/resolve all of the following...</summary>

### Integration Impact

Does this change affect any of these integrations?
- [ ] DfE Analytics platform
- [ ] Legacy imports mappings
- [ ] DWP Find a Job export mappings
- [ ] Publisher ATS API (may require mapping updates or API versioning)

### User Experience & Data Integrity

Could this change impact:
- [ ] Existing subscription alerts (will legacy subscription search
filters break?)
- [ ] Legacy vacancy copying (will copied vacancies fail new
validations?)
- [ ] In-progress drafts for Vacancies or Job Applications
</details>