subscription jobs via global id (#8282)
## Changes in this PR:

Trigger subscription jobs via global id. 

The pattern of passing AR objects to background jobs seems to have been
missed in this project - despite the line in ApplicationJob which
discards jobs with DeserializationErrors (which is thrown when global id
objects cannot be found when the job is run). Creating this here as a
pattern which can be followed, avoiding unnecessary code and tests for
loading AR records from the database during job execution