Allow users to tab through trix buttons on messages (#8260)
Prior to this change users couldn't tab through the formatting buttons
(bold/italic etc) on the messages text box. This change allows them to
tab through them by removing the tab index of -1 which trix adds to
these buttons and adds styling to help user see which button they are
tabbing through, in line with the rest of the service.