Drop body column from action_text_rich_texts (#8335)
## Trello card URL

## Changes in this PR:

This PR drops the body column from the action_text_rich_texts table as
we are now encrypting and storing them in the body_ciphertext column.
This is the third PR to ensure that the action_text_rich_texts table
stores only encrypted data. From this point on there will be no
unencrypted data on the table.

Step 1 is
[here](https://github.com/DFE-Digital/teaching-vacancies/pull/8321),
step 2 is
[here](https://github.com/DFE-Digital/teaching-vacancies/pull/8334).

The final step will be to remove the ignored_column from
config/initializers/action_text.rb and delete the rake task and
associated test used to migrate the data from body to body_ciphertext

## Screenshots of UI changes:

### Before

### After


## Checklists:

### Data & Schema Changes

If this PR modifies data structures or validations, check the following:

- [ ] Adds/removes model validations
- [ ] Adds/removes database fields
- [ ] Modifies Vacancy enumerables (phases, working patterns, job roles,
key stages, etc.)

<details>
<summary>If any of the above options has changed then the author must
check/resolve all of the following...</summary>

### Integration Impact

Does this change affect any of these integrations?
- [ ] DfE Analytics platform
- [ ] Legacy imports mappings
- [ ] DWP Find a Job export mappings
- [ ] Publisher ATS API (may require mapping updates or API versioning)

### User Experience & Data Integrity

Could this change impact:
- [ ] Existing subscription alerts (will legacy subscription search
filters break?)
- [ ] Legacy vacancy copying (will copied vacancies fail new
validations?)
- [ ] In-progress drafts for Vacancies or Job Applications
</details>