Change organisation slug when name changes and keep a history of slug changes (#8358)
## Trello card URL


https://trello.com/c/IfNepRMC/2428-investigate-updating-wythensawe-catholic-academy-trust-urls-to-match-new-name-corpus-christi-catholic-academy-trust

## Changes in this PR:

We had a complaint from a trust that they had changed the name of their
trust but the slug in the URL still referenced their old name.

This PR:

- adds logic to create a new slug when an organisation changes it's
name. Currently the slug is only created when an organisation is
created, and never updated.
- adds history module to friendly_id for organisations. This is already
used on the vacancy model. This tracks all historical slugs in the
friendly_id_slugs table, and automatically redirects old slug URLs to
the current slug which will preserve existing bookmarks/links etc.

Note: We will need to manually force the slug to update because in this
case the name of the school has already changed before we introduced
this logic. To do this I will set organisation.slug = nil, then call
organisation.save causing a new slug to be created for the trust in
question. However going forward this won't be necessary as the slug will
update when an organisation changes it's name. I feel confident making
this change in the console because we import organisations every day and
so it isn't a very high risk change imo.

## Screenshots of UI changes:

### Before

### After


## Checklists:

### Data & Schema Changes

If this PR modifies data structures or validations, check the following:

- [ ] Adds/removes model validations
- [ ] Adds/removes database fields
- [ ] Modifies Vacancy enumerables (phases, working patterns, job roles,
key stages, etc.)

<details>
<summary>If any of the above options has changed then the author must
check/resolve all of the following...</summary>

### Integration Impact

Does this change affect any of these integrations?
- [ ] DfE Analytics platform
- [ ] Legacy imports mappings
- [ ] DWP Find a Job export mappings
- [ ] Publisher ATS API (may require mapping updates or API versioning)

### User Experience & Data Integrity

Could this change impact:
- [ ] Existing subscription alerts (will legacy subscription search
filters break?)
- [ ] Legacy vacancy copying (will copied vacancies fail new
validations?)
- [ ] In-progress drafts for Vacancies or Job Applications
</details>

---------

Co-authored-by: Kyle MacPherson <macphersonkd@gmail.com>