Populate the friendly_id_slugs table (#8360)
## Trello card URL

Related to
https://trello.com/c/IfNepRMC/2428-investigate-updating-wythensawe-catholic-academy-trust-urls-to-match-new-name-corpus-christi-catholic-academy-trust.
PR here: https://github.com/DFE-Digital/teaching-vacancies/pull/8358

## Changes in this PR:

Triggering a slug change did not work as expected because friendly_id
did not account for a conflict with another trust with the same name,
and thus couldn't save the slug it automatically assigned due to the
uniqueness validation on slug in our database.

The reason for this is that the friendly_id gem searches for conflicts
only in the friendly_id_slugs table, ours is empty for organisations.
The rake task in this PR populates the table with all current slugs so
that going forward friendly_id will find conflicts as expected and
create non conflicting slugs where necessary.

## Screenshots of UI changes:

### Before

### After


## Checklists:

### Data & Schema Changes

If this PR modifies data structures or validations, check the following:

- [ ] Adds/removes model validations
- [ ] Adds/removes database fields
- [ ] Modifies Vacancy enumerables (phases, working patterns, job roles,
key stages, etc.)

<details>
<summary>If any of the above options has changed then the author must
check/resolve all of the following...</summary>

### Integration Impact

Does this change affect any of these integrations?
- [ ] DfE Analytics platform
- [ ] Legacy imports mappings
- [ ] DWP Find a Job export mappings
- [ ] Publisher ATS API (may require mapping updates or API versioning)

### User Experience & Data Integrity

Could this change impact:
- [ ] Existing subscription alerts (will legacy subscription search
filters break?)
- [ ] Legacy vacancy copying (will copied vacancies fail new
validations?)
- [ ] In-progress drafts for Vacancies or Job Applications
</details>

Co-authored-by: Kyle MacPherson <macphersonkd@gmail.com>