Personal statement richtext (#8347)
## Trello card URL


https://trello.com/c/YRGDGRao/2335-spike-update-application-personal-statement-page-with-the-comms-component

## Changes in this PR:

This PR is the first step in adding rich text to personal statements.

This PR adds a rich text form to the personal details page. It is
similar to the form on the messages page but does not include the
hyperlink or attachment options. It also adds a migration which will
copy data from the personal_details column on the job_application table
to the action_text_rich_texts table for existing job applications.

Once this PR is merged, jobseekers will be using rich text for personal
statements on new job applications. Their previously job applications
will still display the existing personal statement even prior to the
migration. Once the migration happens all job applications with personal
statements will have them transferred over to the rich texts table.

The next and final step of this work will be to:
- remove all the fall back logic which displays the old plain text
personal statements (which won't be needed as when the migration is
successfully run all job applications will have their personal
statements stored in the action text rich texts table)
- remove the rake task and job. 
- drop the personal_statement column on job applications

I would also like to investigate changing the variable name `content` to
`personal_statement` to see if we can reference the entry in the rich
texts table in a more intuitive way going forward. We can't do this
until the migration has been run and the personal_statement column has
been dropped from the job application table.

Note: I have also fixed a bug in this PR which was causing users to not
be able to apply for religious jobs after using a baptism certificate
previously. [This
bug](https://teaching-vacancies.sentry.io/issues/7084973182/?environment=review&query=is%3Aunresolved&referrer=issue-stream)
crashed the page when users applied for another religious vacancy
because the logic to attach the baptism certificate when prefilling the
job application was broken. When testing the prefilling of
personal_statements from the new action_text_rich_texts table I added
tests for this part of the codebase that was previously untested, in
doing so I came across this bug and just added it to this PR, but it's
not really related to this change so I can raise new PR if necessary.

## Screenshots of UI changes:

### Before

### After


## Checklists:

### Data & Schema Changes

If this PR modifies data structures or validations, check the following:

- [ ] Adds/removes model validations
- [ ] Adds/removes database fields
- [ ] Modifies Vacancy enumerables (phases, working patterns, job roles,
key stages, etc.)

<details>
<summary>If any of the above options has changed then the author must
check/resolve all of the following...</summary>

### Integration Impact

Does this change affect any of these integrations?
- [ ] DfE Analytics platform
- [ ] Legacy imports mappings
- [ ] DWP Find a Job export mappings
- [ ] Publisher ATS API (may require mapping updates or API versioning)

### User Experience & Data Integrity

Could this change impact:
- [ ] Existing subscription alerts (will legacy subscription search
filters break?)
- [ ] Legacy vacancy copying (will copied vacancies fail new
validations?)
- [ ] In-progress drafts for Vacancies or Job Applications
</details>

---------

Co-authored-by: Kyle MacPherson <macphersonkd@gmail.com>