Personal statement tidy up (#8371)
## Trello card URL

## Changes in this PR:
This is a follow up to the [initial
PR](https://github.com/DFE-Digital/teaching-vacancies/pull/8347) to
convert personal statements to rich text.

This PR ignores the personal_statement column on the job applications
table which we need to do before dropping the column in the next PR. It
also removes references to job_application.personal_statement throughout
the codebase and removes the rake tasks and associated tests that we no
longer need.

## Screenshots of UI changes:

### Before

### After


## Checklists:

### Data & Schema Changes

If this PR modifies data structures or validations, check the following:

- [ ] Adds/removes model validations
- [ ] Adds/removes database fields
- [ ] Modifies Vacancy enumerables (phases, working patterns, job roles,
key stages, etc.)

<details>
<summary>If any of the above options has changed then the author must
check/resolve all of the following...</summary>

### Integration Impact

Does this change affect any of these integrations?
- [ ] DfE Analytics platform
- [ ] Legacy imports mappings
- [ ] DWP Find a Job export mappings
- [ ] Publisher ATS API (may require mapping updates or API versioning)

### User Experience & Data Integrity

Could this change impact:
- [ ] Existing subscription alerts (will legacy subscription search
filters break?)
- [ ] Legacy vacancy copying (will copied vacancies fail new
validations?)
- [ ] In-progress drafts for Vacancies or Job Applications
</details>