Bugfix: Notification loading before authentication
When the jobseekers try to access their notifications without being signed in,
the code tries to load the notifications for the user, but the user is
still not set and it raises an exception.

Fixed it by ensuring the authentication is prepended before any other
before action hook. And add some request specs to cover the behaviour.